authentication added

+++ RolePlayingGame/Services/CharacterService/CharacterService.cs
@@ -3,33 +3,45 @@ using Microsoft.EntityFrameworkCore;
 using RolePlayingGame.Data;
 using RolePlayingGame.Dtos.Character;
 using RolePlayingGame.Models;
+using System.Security.Claims;
 
 namespace RolePlayingGame.Services.CharacterService
 {
     public class CharacterService : ICharacterService
     {
-        private static List<Character> Characters = new List<Character>
-        {
-            new Character(),
-            new Character{ Id = 1, Name = "Samuel", Class = RpgClass.Cleric}
-        };
         private readonly IMapper _mapper;
         private readonly DataContext _dbContext;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _dbContext = context;
+            _httpContextAccessor = httpContextAccessor;
+        }
+
+        private int GetUserId()
+        {
+            return int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+        }
+
+        private async Task<List<GetCharacterDto>> GetCharactersForCurrentUser()
+        {
+            return await _dbContext.Characters
+                    .Where(c => c.User.Id == GetUserId())
+                    .Select(c => _mapper.Map<GetCharacterDto>(c))
+                    .ToListAsync();
         }
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddNewCharacter(AddCharacterDto newCharacter)
         {
             var character = _mapper.Map<Character>(newCharacter);
+            character.User = await _dbContext.Users.FirstOrDefaultAsync(x => x.Id == GetUserId());
             await _dbContext.Characters.AddAsync(character);
             await _dbContext.SaveChangesAsync();
             var response = new ServiceResponse<List<GetCharacterDto>>
             {
-                Data = (_dbContext.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList()
+                Data = await GetCharactersForCurrentUser()
             };
             return response;
         }
@@ -39,11 +51,16 @@ namespace RolePlayingGame.Services.CharacterService
             var response = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                var character = await _dbContext.Characters.FirstAsync(c => c.Id == id);
+                var character = await _dbContext.Characters
+                    .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if (character == null)
+                {
+                    throw new KeyNotFoundException();
+                }
                 _dbContext.Characters.Remove(character);
                 await _dbContext.SaveChangesAsync();
 
-                response.Data = (_dbContext.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                response.Data = await GetCharactersForCurrentUser();
             }
             catch (Exception e)
             {
@@ -55,10 +72,9 @@ namespace RolePlayingGame.Services.CharacterService
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
-            var dbCharacters = await _dbContext.Characters.ToListAsync();
             var response = new ServiceResponse<List<GetCharacterDto>>
             {
-                Data = (dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList()
+                Data = await GetCharactersForCurrentUser()
             };
             return response;
         }
@@ -68,7 +84,8 @@ namespace RolePlayingGame.Services.CharacterService
             var response = new ServiceResponse<GetCharacterDto>();
             try
             {
-                var character = await _dbContext.Characters.FirstOrDefaultAsync(c => c.Id == id);
+                var character = await _dbContext.Characters
+                    .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
                 if (character == null)
                 {
                     throw new KeyNotFoundException();
